requirements.txt
pandas
scikit-learn
pickle-mixin
numpy
pyyaml
matplotlib
seaborn
statsmodels
squarify
jinja2